integrate SPM package processing in graph command
